pyproject.toml
[project]
name = "regulation-kg-rag"
version = "0.1.0"
description = "Knowledge Graph based Retrieval Augmented Generation for Risk Assessment Regulation in Korea"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "attrs==25.3.0",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.2",
    "colorama==0.4.6",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "fire==0.7.0",
    "frozenlist==1.7.0",
    "fsspec==2024.12.0",
    "greenlet==3.2.3",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "idna==3.10",
    "interchange==2021.0.4",
    "jiter==0.10.0",
    "json-repair==0.44.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==0.3.27",
    "langchain-community==0.3.27",
    "langchain-core==0.3.72",
    "langchain-neo4j==0.5.0",
    "langchain-openai==0.3.28",
    "langchain-text-splitters==0.3.9",
    "langsmith==0.4.12",
    "marshmallow==3.26.1",
    "monotonic==1.6",
    "multidict==6.6.3",
    "mypy-extensions==1.1.0",
    "neo4j==5.28.2",
    "neo4j-graphrag==1.9.0",
    "numpy==2.3.2",
    "openai==1.99.1",
    "orjson==3.11.1",
    "packaging==25.0",
    "pansi==2024.11.0",
    "pillow==11.3.0",
    "propcache==0.3.2",
    "py2neo==2021.2.4",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pygments==2.19.2",
    "pypdf==5.9.0",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "regex==2025.7.34",
    "requests==2.32.4",
    "requests-toolbelt==1.0.0",
    "scipy==1.16.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.42",
    "tenacity==9.1.2",
    "termcolor==3.1.0",
    "tiktoken==0.10.0",
    "tqdm==4.67.1",
    "types-pyyaml==6.0.12.20250516",
    "typing-extensions==4.14.1",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "urllib3==2.5.0",
    "xmltodict==0.14.2",
    "yarl==1.20.1",
    "zstandard==0.23.0",
]
